requirements.txt
flask==1.1.2
flatten-json==0.1.13
numpy==1.18.5
pandas==1.0.5
python-dotenv==0.18.0
requests==2.24.0





